# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Customer job requests and dispatch state
  - name: partners
    description: Trade partner onboarding and availability
  - name: admin
    description: Admin dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job requests (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Job requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a customer job request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Request received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get public status for a job using access token
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/JobStatusToken"
      responses:
        "200":
          description: Public job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJobStatus"
        "401":
          description: Missing or invalid token
        "404":
          description: Job not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Authentication required
        "403":
          description: Forbidden

  /jobs/{id}/intake:
    patch:
      operationId: correctJobIntake
      tags: [jobs]
      summary: Correct customer-confirmed job details
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/JobStatusToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobIntakeCorrection"
      responses:
        "200":
          description: Corrected request status and current intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJobStatus"
        "401":
          description: Missing or invalid token
        "404":
          description: Job not found

  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List trade partners (admin gets all; partner gets self)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trade partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
    post:
      operationId: createPartner
      tags: [partners]
      summary: Submit a trade partner application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"

  /partners/{id}/availability:
    patch:
      operationId: updatePartnerAvailability
      tags: [partners]
      summary: Update trade partner availability
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "401":
          description: Authentication required
        "403":
          description: Forbidden

  /dispatches/{id}/decision:
    patch:
      operationId: decideDispatch
      tags: [jobs]
      summary: Accept or decline a dispatched opportunity
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchDecisionInput"
      responses:
        "200":
          description: Updated dispatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
        "401":
          description: Authentication required
        "403":
          description: Forbidden

  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get admin dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          description: Authentication required
        "403":
          description: Forbidden

  /admin/jobs-awaiting-dispatch:
    get:
      operationId: listJobsAwaitingDispatch
      tags: [admin]
      summary: List jobs awaiting dispatch
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Jobs awaiting dispatch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  /admin/jobs/{id}/partner-recommendations:
    get:
      operationId: getPartnerRecommendations
      tags: [admin]
      summary: List deterministic partner recommendations without creating offers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Ranked partner recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerRecommendation"

  /admin/approved-partners:
    get:
      operationId: listApprovedPartners
      tags: [admin]
      summary: List approved partners eligible for dispatch
      security:
        - bearerAuth: []
      parameters:
        - name: trade
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Approved partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"

  /admin/dispatch-offers:
    get:
      operationId: listDispatchOffers
      tags: [admin]
      summary: List dispatch offer history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dispatch offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DispatchOffer"
    post:
      operationId: createDispatchOffer
      tags: [admin]
      summary: Create a dispatch offer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchOfferInput"
      responses:
        "201":
          description: Dispatch offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchOffer"

  /partner/offers:
    get:
      operationId: listPartnerOffers
      tags: [partners]
      summary: List offers for the authenticated partner
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Partner opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerOffer"

  /admin/dispatch-offers/{id}/expire:
    patch:
      operationId: expireDispatchOffer
      tags: [admin]
      summary: Expire a pending dispatch offer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Expired dispatch offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
    JobStatusToken:
      name: token
      in: query
      required: true
      schema:
        type: string
        minLength: 16

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Job:
      type: object
      required: [id, reference, description, trade, suburb, postcode, urgency, preferredTime, status, createdAt, assessment]
      properties:
        id: { type: number }
        reference: { type: string }
        description: { type: string }
        trade: { type: string }
        suburb: { type: string }
        postcode: { type: string }
        urgency: { type: string }
        preferredTime: { type: string }
        status: { type: string }
        customerName: { type: string }
        customerPhone: { type: ["string", "null"] }
        customerEmail: { type: ["string", "null"] }
        createdAt: { type: string }
        images:
          type: array
          items: { type: string }
        assessment:
          anyOf:
            - $ref: "#/components/schemas/JobAssessment"
            - type: "null"

    CreateJobResponse:
      type: object
      required: [id, reference, status, createdAt, statusAccessToken, statusAccessUrl]
      properties:
        id: { type: number }
        reference: { type: string }
        status: { type: string }
        createdAt: { type: string }
        statusAccessToken: { type: string }
        statusAccessUrl: { type: string }

    PublicJobStatus:
      type: object
      required: [reference, status, createdAt, updatedAt, intake]
      properties:
        reference: { type: string }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        intake:
          $ref: "#/components/schemas/CustomerConfirmedIntake"
        assessment:
          anyOf:
            - $ref: "#/components/schemas/JobAssessment"
            - type: "null"

    JobInput:
      type: object
      required: [description, trade, suburb, postcode, urgency, preferredTime, customerName]
      properties:
        description: { type: string, minLength: 4 }
        trade: { type: string }
        suburb: { type: string }
        postcode: { type: string }
        urgency: { type: string }
        preferredTime: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string }
        images:
          type: array
          items: { type: string }

    JobUpdate:
      type: object
      properties:
        status: { type: string }

    JobIntakeCorrection:
      type: object
      required: [description, trade, suburb, postcode, urgency, preferredTime, customerName]
      properties:
        description: { type: string, minLength: 4 }
        trade: { type: string, minLength: 1 }
        suburb: { type: string, minLength: 1 }
        postcode: { type: string, minLength: 1 }
        urgency: { type: string, minLength: 1 }
        preferredTime: { type: string, minLength: 1 }
        customerName: { type: string, minLength: 1 }
        customerPhone: { type: string }
        customerEmail: { type: string }

    CustomerConfirmedIntake:
      type: object
      required: [description, trade, suburb, postcode, urgency, preferredTime, customerName]
      properties:
        description: { type: string }
        trade: { type: string }
        suburb: { type: string }
        postcode: { type: string }
        urgency: { type: string }
        preferredTime: { type: string }
        customerName: { type: string }
        customerPhone: { type: ["string", "null"] }
        customerEmail: { type: ["string", "null"] }

    JobAssessment:
      type: object
      required: [outcome, provider, model, safetyCodes, assessment, createdAt]
      properties:
        outcome: { type: string }
        provider: { type: string }
        model: { type: ["string", "null"] }
        safetyCodes:
          type: array
          items: { type: string }
        assessment:
          $ref: "#/components/schemas/JobIntakeAssessmentDraft"
        createdAt: { type: string }

    TradeClassification:
      type: string
      enum: [plumbing, electrical, heating_cooling, unsure]

    UrgencyClassification:
      type: string
      enum: [not_urgent, soon, today, emergency, unsure]

    PhotoContext:
      type: object
      additionalProperties: false
      required: [provided, count]
      properties:
        provided: { type: boolean }
        count: { type: number, minimum: 0 }

    JobIntakeAssessmentDraft:
      type: object
      additionalProperties: false
      required:
        - tradeClassification
        - urgencyClassification
        - suburb
        - postcode
        - preferredAttendanceTime
        - neutralProblemSummary
        - equipment
        - brand
        - model
        - photoContext
        - confidence
        - codes
      properties:
        tradeClassification:
          $ref: "#/components/schemas/TradeClassification"
        urgencyClassification:
          $ref: "#/components/schemas/UrgencyClassification"
        suburb: { type: ["string", "null"], maxLength: 120 }
        postcode: { type: ["string", "null"], maxLength: 16 }
        preferredAttendanceTime: { type: ["string", "null"], maxLength: 160 }
        neutralProblemSummary: { type: ["string", "null"], maxLength: 500 }
        equipment: { type: ["string", "null"], maxLength: 120 }
        brand: { type: ["string", "null"], maxLength: 120 }
        model: { type: ["string", "null"], maxLength: 120 }
        photoContext:
          $ref: "#/components/schemas/PhotoContext"
        confidence:
          type: string
          enum: [low, medium, high]
        codes:
          type: array
          maxItems: 3
          items:
            type: string
            enum: [ROUTING_REVIEW, URGENCY_REVIEW, MANUAL_REVIEW_REQUIRED]

    Partner:
      type: object
      required: [id, businessName, contactName, trade, suburbs, availability, status]
      properties:
        id: { type: number }
        businessName: { type: string }
        contactName: { type: string }
        abn: { type: ["string", "null"] }
        trade: { type: string }
        suburbs:
          type: array
          items: { type: string }
        radiusKm: { type: number }
        availability: { type: boolean }
        status: { type: string }
        services:
          type: array
          items: { type: string }
        emergencyJobs: { type: boolean }

    PartnerInput:
      type: object
      required: [businessName, contactName, trade, mobile, email, suburbs, radiusKm]
      properties:
        businessName: { type: string }
        contactName: { type: string }
        abn: { type: string }
        trade: { type: string }
        licence: { type: string }
        mobile: { type: string }
        email: { type: string }
        suburbs:
          type: array
          items: { type: string }
        radiusKm: { type: number }
        services:
          type: array
          items: { type: string }
        emergencyJobs: { type: boolean }

    AvailabilityInput:
      type: object
      required: [availability]
      properties:
        availability: { type: boolean }

    Dispatch:
      type: object
      required: [id, jobId, businessId, decision]
      properties:
        id: { type: number }
        jobId: { type: number }
        businessId: { type: number }
        decision: { type: string }
        offeredAt: { type: string }
        respondedAt: { type: ["string", "null"] }

    DispatchDecisionInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string }

    DispatchOfferInput:
      type: object
      required: [jobId, partnerId, expiresAt]
      properties:
        jobId: { type: number }
        partnerId: { type: number }
        expiresAt: { type: string, format: date-time }

    DispatchOffer:
      type: object
      required: [id, jobId, partnerId, state, offeredAt]
      properties:
        id: { type: number }
        jobId: { type: number }
        partnerId: { type: number }
        state: { type: string }
        offeredAt: { type: string }
        respondedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }

    PartnerOffer:
      allOf:
        - $ref: "#/components/schemas/DispatchOffer"
        - type: object
          properties:
            job:
              $ref: "#/components/schemas/PartnerOfferJob"

    PartnerOfferJob:
      type: object
      required: [reference, trade, suburb, postcode, urgency, preferredTime, description, createdAt, updatedAt]
      properties:
        reference: { type: string }
        trade: { type: string }
        suburb: { type: string }
        postcode: { type: string }
        urgency: { type: string }
        preferredTime: { type: string }
        description: { type: string }
        customerName: { type: ["string", "null"] }
        customerPhone: { type: ["string", "null"] }
        customerEmail: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }

    PartnerRecommendation:
      type: object
      required: [partnerId, score, eligible, codes, disqualifications]
      properties:
        partnerId: { type: number }
        score: { type: number }
        eligible: { type: boolean }
        codes:
          type: array
          items: { type: string }
        disqualifications:
          type: array
          items: { type: string }

    AdminSummary:
      type: object
      required: [newRequests, awaitingDispatch, tradieApplications, approvedTradies, availableTradies, sentOpportunities, acceptedJobs, declinedJobs, completedJobs]
      properties:
        newRequests: { type: number }
        awaitingDispatch: { type: number }
        tradieApplications: { type: number }
        approvedTradies: { type: number }
        availableTradies: { type: number }
        sentOpportunities: { type: number }
        acceptedJobs: { type: number }
        declinedJobs: { type: number }
        completedJobs: { type: number }