# lib/api-zod/src/generated/types/customerConfirmedIntake.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerConfirmedIntake {
  description: string;
  trade: string;
  suburb: string;
  postcode: string;
  urgency: string;
  preferredTime: string;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerEmail?: string | null;
}