# lib/api-zod/src/generated/types/jobAssessment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobIntakeAssessmentDraft } from './jobIntakeAssessmentDraft';

export interface JobAssessment {
  outcome: string;
  provider: string;
  /** @nullable */
  model: string | null;
  safetyCodes: string[];
  assessment: JobIntakeAssessmentDraft;
  createdAt: string;
}