# lib/api-zod/src/generated/types/jobIntakeAssessmentDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobIntakeAssessmentDraftCodesItem } from './jobIntakeAssessmentDraftCodesItem';
import type { JobIntakeAssessmentDraftConfidence } from './jobIntakeAssessmentDraftConfidence';
import type { PhotoContext } from './photoContext';
import type { TradeClassification } from './tradeClassification';
import type { UrgencyClassification } from './urgencyClassification';

export interface JobIntakeAssessmentDraft {
  tradeClassification: TradeClassification;
  urgencyClassification: UrgencyClassification;
  /**
     * @maxLength 120
     * @nullable
     */
  suburb: string | null;
  /**
     * @maxLength 16
     * @nullable
     */
  postcode: string | null;
  /**
     * @maxLength 160
     * @nullable
     */
  preferredAttendanceTime: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  neutralProblemSummary: string | null;
  /**
     * @maxLength 120
     * @nullable
     */
  equipment: string | null;
  /**
     * @maxLength 120
     * @nullable
     */
  brand: string | null;
  /**
     * @maxLength 120
     * @nullable
     */
  model: string | null;
  photoContext: PhotoContext;
  confidence: JobIntakeAssessmentDraftConfidence;
  /** @maxItems 3 */
  codes: JobIntakeAssessmentDraftCodesItem[];
}