# lib/api-zod/src/generated/types/jobIntakeCorrection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobIntakeCorrection {
  /** @minLength 4 */
  description: string;
  /** @minLength 1 */
  trade: string;
  /** @minLength 1 */
  suburb: string;
  /** @minLength 1 */
  postcode: string;
  /** @minLength 1 */
  urgency: string;
  /** @minLength 1 */
  preferredTime: string;
  /** @minLength 1 */
  customerName: string;
  customerPhone?: string;
  customerEmail?: string;
}