# lib/api-zod/src/generated/types/partnerRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerRecommendation {
  partnerId: number;
  score: number;
  eligible: boolean;
  codes: string[];
  disqualifications: string[];
}