# lib/api-zod/src/generated/types/tradeClassification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TradeClassification = typeof TradeClassification[keyof typeof TradeClassification];


export const TradeClassification = {
  plumbing: 'plumbing',
  electrical: 'electrical',
  heating_cooling: 'heating_cooling',
  unsure: 'unsure',
} as const;